resolved issue brougth up by @caw
